Disconnect resizeMinMaxOpId on disable
resizeMinMaxOpId is a signal handler from the global.window_manager
object, not the global.display one. This patch fixes the disable
function disconnecting this signal handler from the corresponding
object.

This will fix a problem that appears when the user disables the
extension but the wobbly effect for maximize and minize is still there
until the next shell reboot.